add a test for formatData, move EventsPerHour to separate file to make testing easier
